requirements.txt
flask
flask-pymongo
flask-bower
python-dotenv
numpy
werkzeug
jinja2
mongoengine
itsdangerous
click
pandas
numpy
requests
sklearn
scipy
Cython
git+git://github.com/khrapovs/skewstudent
git+git://github.com/khrapovs/bekk
